lsp: Removed ray-x/navigator.
Removed ray-x/navigator, because its too overpowering. Yes it's a good
plugin, but i wanted it for code display, but it takes more roles than
that that i like the stock behaviour for. The simple highlight for vim,
when there is an error is better than highlighting it in purple(my
color scheme)